Namespace isic-partial files by pid
This prevents the tool from trying to delete partial files that are
created by other (possibly concurrent) invocations of isic-cli which has
been seen in the wild. It doesn't imply support for concurrent
executions of isic-cli with the same output directory, but it at least
prevents it from trying to tamper with files outside of its custody.